open mergeTemplate function, make media listener be able to get exist state
